refactor(connect): move evolu method definitions to corresponding folder
