+++ packages/connect-common/src/types/api/evolu/evoluGetDelegatedIdentityKey.ts
@@ -1,6 +1,6 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { CommonParams, Params, Response } from '../params';
+import type { CommonParams, Params, Response } from '../../params';
 
 export declare function evoluGetDelegatedIdentityKey(
     params: Params<CommonParams>,
