+++ packages/connect-common/src/types/api/evolu/evoluGetNode.ts
@@ -1,5 +1,5 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { Params, Response } from '../params';
+import type { Params, Response } from '../../params';
 
 export declare function evoluGetNode(params: Params<PROTO.EvoluGetNode>): Response<PROTO.EvoluNode>;
