+++ packages/connect-common/src/types/api/evolu/evoluSignRegistrationRequest.ts
@@ -1,6 +1,6 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { Params, Response } from '../params';
+import type { Params, Response } from '../../params';
 
 export declare function evoluSignRegistrationRequest(
     params: Params<PROTO.EvoluSignRegistrationRequest>,
